feat: Add custom journey editing and journey management pages
- Implemented CustomJourneyEditor for editing custom journeys with drag-and-drop functionality for arcs.
- Created JourneyPage to display current journey progress, including options to complete, skip, or delete journeys.
- Added StartJourneyPage for users to initiate their journey with options for custom and prebuilt journeys.
- Developed LoginPage and RegisterPage for user authentication with error handling.
- Introduced MeditationDayPage for displaying daily meditation content and managing notes.
- Updated NotesPage layout for improved user experience.
- Minor adjustments to Tailwind CSS configuration and TypeScript settings.
