+++ frontend/jest.config.cjs
@@ -5,11 +5,9 @@ module.exports = {
     '^@/(.*)$': '<rootDir>/src/$1',
   },
   testMatch: ['**/?(*.)+(test).[jt]s?(x)', '**/*.test.ts', '**/*.test.tsx'],
-  extensionsToTreatAsEsm: ['.ts', '.tsx'], // Treat TypeScript files as ESM
   transform: {
     '^.+\\.tsx?$': ['ts-jest', { tsconfig: 'tsconfig.json' }],
-    '^.+\\.jsx?$': 'babel-jest', // Transform JavaScript files with Babel
   },
-  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
-  extensionsToTreatAsEsm: ['.ts', '.tsx'],
+  extensionsToTreatAsEsm: ['.ts', '.tsx'], // Treat TypeScript files as ESM
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'], // Ensure this is a valid path
 };
