+++ frontend/tailwind.config.js
@@ -1,4 +1,3 @@
-// tailwind.config.js
 import plugin from 'tailwindcss/plugin';
 import lineClamp from '@tailwindcss/line-clamp';
