+++ frontend/tsconfig.json
@@ -33,6 +33,6 @@
   "include": [
     "src/**/*",
     "vite.config.ts",
-    "tailwind.config.js"
+    "tailwind.config.js",
   ],
 }
